tests/wpt/web-platform-tests/css/css-view-transitions/no-containment-on-new-element-mid-transition.html

<!DOCTYPE html>
<html>
<title>View transitions: transition skipped if no containment on new element after animation started</title>
<link rel="help" href="https://www.w3.org/TR/css-view-transitions-1/">
<link rel="author" href="mailto:khushalsagar@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  width: 100px;
  height: 100px;
  background: blue;
  view-transition-name: target;
  contain: paint;
}

html::view-transition-group(target) {
  animation-duration: 300s;
}
</style>

<div id=first></div>

<script>
promise_test(async t => {
  assert_implements(document.startViewTransition, "Missing document.startViewTransition");
  return new Promise(async (resolve, reject) => {
    let transition = document.startViewTransition();
    await transition.domUpdated;
    await transition.ready;

    transition.finished.then(resolve, reject);
    first.style.contain = "none";
  });
}, "new element becoming uncontained should skip the transition");
</script>
